[REFACTOR] separate constructor tests for invoice controller
